+++ src/main/kotlin/view/CanvasView.kt
@@ -276,77 +276,3 @@ fun Canvas(viewModel: CanvasViewModel) {
     }
 
 }
-
-enum class StorageType {
-    FILE,
-    NEO4J,
-    SQLITE
-}
-
-@Composable
-fun CustomRadioGroup(
-    options: List<String>,
-    selectedOption: String,
-    onOptionSelected: (String) -> Unit
-) {
-    Column {
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .background(Color.Magenta)
-                    .padding(8.dp)
-                    .weight(1f)
-            ) {
-                Text("NEO4J", color = Color.White)
-            }
-            Spacer(modifier = Modifier.width(8.dp))
-            Button(
-                onClick = { onOptionSelected(StorageType.NEO4J.name) }
-            ) {
-                Text("Select")
-            }
-        }
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .background(Color.Magenta)
-                    .padding(8.dp)
-                    .weight(1f)
-            ) {
-                Text("FILE", color = Color.White)
-            }
-            Spacer(modifier = Modifier.width(8.dp))
-            Button(
-                onClick = { onOptionSelected(StorageType.FILE.name) }
-            ) {
-                Text("Select")
-            }
-        }
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Row(
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .background(Color.Magenta)
-                    .padding(8.dp)
-                    .weight(1f)
-            ) {
-                Text("SQLITE", color = Color.White)
-            }
-            Spacer(modifier = Modifier.width(8.dp))
-            Button(
-                onClick = { onOptionSelected(StorageType.SQLITE.name) }
-            ) {
-                Text("Select")
-            }
-        }
-    }
-}
