+++ src/main/kotlin/view/StorageType.kt
@@ -1,0 +1,7 @@
+package view
+
+enum class StorageType {
+    FILE,
+    NEO4J,
+    SQLITE
+}
